Dashboards page - infrastructure and test (#394)
* Dashboards page - infrastructure and test

This adds the basic infrastructure to interact with the elements in the
dashboards page. It is still WIP and currently includes support for:

1) Reading the inventory and statuses of the inventory items (hosts and
   nodes, for example, and how many of them are up and how many down)
2) Reading the events (such as pod started, pod stopped etc...)
3) Reading the alerts (such as storage space low etc...)
4) A small test is included, which tests that the nodes and hosts in
   the inventory match what you get from the CLI

* Lint fixes

* Fixes for the comments in the first review

- identifying the rows in the inventory card by their name and not hard-
  coded index
- replacing some for loops with forEach (but not all of them - because
  you can't break or return from a forEach loop)
- utility function to check if an element exists and read the text from
  it

* Refactoring to use beforeAll

Adding a beforeAll() method to handle the browsing to the dashboards
page. Also changed the suite name to "baremetalSmokeTests", as agreed,
to be able to run the fast tests on the gate and run the e2e tests
separately in the future.

* Adding block comments

Documenting the elements and functions with block comments

* Added function expectCounters

Created expectCounters(rowNumber, upCounterExpected, downCounterExpected, label)
in order to reduce the amount of copy-pasted code.